Displacement sampling basics working and demo'd

+++ crest/Assets/Crest/Crest/Shaders/OceanLODData.hlsl
@@ -53,7 +53,7 @@ float2 LD_WorldToUV(in float2 i_samplePos, in float2 i_centerPos, in float i_res
 	return (i_samplePos - i_centerPos) / (i_texelSize * i_res) + 0.5;
 }
 
-float3 WorldToUV(in float2 i_samplePos, in float i_sliceIndex) {
+float3 WorldToUV(in float2 i_samplePos, in uint i_sliceIndex) {
 	const float2 result = LD_WorldToUV(
 		i_samplePos,
 		_LD_Pos_Scale[i_sliceIndex].xy,
@@ -63,7 +63,7 @@ float3 WorldToUV(in float2 i_samplePos, in float i_sliceIndex) {
 	return float3(result, i_sliceIndex);
 }
 
-float3 WorldToUV_BiggerLod(in float2 i_samplePos, in float i_sliceIndex_BiggerLod) {
+float3 WorldToUV_BiggerLod(in float2 i_samplePos, in uint i_sliceIndex_BiggerLod) {
 	const float2 result = LD_WorldToUV(
 		i_samplePos, _LD_Pos_Scale[i_sliceIndex_BiggerLod].xy,
 		_LD_Params[i_sliceIndex_BiggerLod].y,
@@ -72,7 +72,7 @@ float3 WorldToUV_BiggerLod(in float2 i_samplePos, in float i_sliceIndex_BiggerLo
 	return float3(result, i_sliceIndex_BiggerLod);
 }
 
-float3 WorldToUV_Source(in float2 i_samplePos, in float i_sliceIndex_Source) {
+float3 WorldToUV_Source(in float2 i_samplePos, in uint i_sliceIndex_Source) {
 	const float2 result = LD_WorldToUV(
 		i_samplePos,
 		_LD_Pos_Scale_Source[i_sliceIndex_Source].xy,
